+++ tests/CSharpLanguageServer.Tests/TypeDefinitionTests.fs
@@ -26,7 +26,7 @@ let ``test textDocument/typeDefinition works`` () =
         |> Uri
         |> string
         |> _.Substring("file:///".Length)
-        |> sprintf "csharp:///%s/Project/Project.csproj?symbol=System.String&view=source"
+        |> sprintf "csharp:/%s/Project/Project.csproj?symbol=System.String&view=source"
 
     match typeDefinition0 with
     | Some(U2.C1(U2.C2 ls)) ->